pyproject.toml
[project]
name = "aws-sidekick"
version = "0.1.0"
description = "AWS Sidekick"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "structlog==24.4.0",
    "aiosqlite>=0.19.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anthropic==0.55.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "aws-requests-auth==0.4.3",
    "blinker==1.9.0",
    "boto3==1.38.45",
    "botocore==1.38.45",
    "cachetools==6.1.0",
    "certifi==2025.6.15",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "diagrams==0.24.4",
    "dill==0.4.0",
    "distlib==0.3.9",
    "distro==1.9.0",
    "docstring-parser==0.16",
    "exceptiongroup==1.3.0",
    "filelock==3.18.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "graphviz==0.20.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "identify==2.6.12",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jmespath==1.0.1",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mcp==1.10.1",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "narwhals==1.44.0",
    "nodeenv==1.9.1",
    "numpy==2.2.6",
    "opentelemetry-api==1.34.1",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-semantic-conventions==0.55b1",
    "packaging==25.0",
    "pandas==2.3.0",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "pre-commit==4.2.0",
    "prompt-toolkit==3.0.51",
    "protobuf==6.31.1",
    "pyarrow==20.0.0",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.4",
    "rich==14.0.0",
    "rpds-py==0.25.1",
    "s3transfer==0.13.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "slack-bolt==1.23.0",
    "slack-sdk==3.35.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sse-starlette==2.3.6",
    "starlette==0.46.2",
    "strands-agents[openai]==0.1.9",
    "strands-agents-tools==0.1.7",
    "streamlit==1.46.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tornado==6.5.1",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "virtualenv==20.31.2",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "wheel==0.45.1",
    "zipp==3.23.0",
]
